src/requirements.txt
numpy==1.26.4
torch
faiss-cpu
sentence-transformers==2.2.2
transformers==4.30.2
tokenizers==0.14.1
huggingface_hub==0.16.4
